apps/api/pyproject.toml
[project]
name = "battlescope-api"
version = "0.1.0"
description = "BattleScope API — FastAPI + LangGraph competitive intelligence agent"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.6.0",
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langsmith>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
battlescope_api = ["prompts/*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]
